backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Environment
python-decouple==3.8
python-dotenv==1.0.1

# Image handling
Pillow==10.2.0

# AWS/Supabase Storage
boto3==1.34.34
supabase==2.3.4

# AI/LLM Integration
openai==1.10.0  # Compatible con Grok API
requests==2.31.0

# Notifications
twilio==8.11.1

# Utilities
cryptography==42.0.2
django-filter==24.1
django-money==3.4.1

# Testing
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Dev tools
black==24.1.1
flake8==7.0.0
